refactor(ui): extract team-sync/render/sharing-review.ts
Move renderSyncSharingReview (and the private openFeedSharingReview helper
it routes to) into its own module. team-sync.ts now re-exports the public
entry so the call site in index.ts stays on the same import path.
